factor regtest mode logic into module
